Separate generated files from examples to fix failing includes
Running the Collector extension on the family directory
`examples` lead to failing includes. It seems not to
be possible to scan this directory, as it is already
known to Antora. Therefore I moved all generated files
to be included to the new directory `generated`.
to ensure Antora will be able to include the
files imported by the Collector extension.